+++ libsolidity/analysis/StaticAnalyzer.h
@@ -65,6 +65,7 @@ private:
 	bool visit(FunctionDefinition const& _function) override;
 	void endVisit(FunctionDefinition const& _function) override;
 
+	bool visit(ForStatement const& _forStatement) override;
 	bool visit(ExpressionStatement const& _statement) override;
 	bool visit(VariableDeclaration const& _variable) override;
 	bool visit(Identifier const& _identifier) override;