+++ test/libsolidity/syntaxTests/uninitialized/for_init_expression_empty.sol
@@ -1,0 +1,10 @@
+contract C {
+    function foo() pure public returns (int) {
+        for (; true;) {
+            int i = 42;
+            return i + 1;
+        }
+        return 0;
+    }
+}
+// ----